requirements.txt
affine==2.4.0
anyio==4.11.0
APScheduler==3.11.1
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
contextily==1.6.2
contourpy==1.3.3
cycler==0.12.1
earthengine-api==1.6.14
fonttools==4.60.1
geographiclib==2.1
geopy==2.4.1
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
mercantile==1.2.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
rasterio==1.4.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
statsmodels==0.14.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
xyzservices==2025.10.0
